Persistent store of current range
This change reads and writes to the RTC registers the current day WeatherSummary information.  This mitigates the problem where the 'forecast' contains less information as the day goes on and results in ranges which are more and more narrow.

This change also allows the current observed temperature to influence the current day range.  This is important since the forecast is estimated and the current observed data is clearly better.