Added general time indicator
In the prior version I was showing the full time as HH:MM which was a little confusing because when you read the display it was easy to interpet this as the current time rather than the time that the display was last updated.

This change makes the time less precise by just indicating the hour.